montera/montera-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
solana==0.30.1
base58==2.1.1
python-dotenv==1.0.1
cryptography==42.0.5
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pymongo==4.7.2
motor==3.4.0
aiofiles==23.2.1
pycares==4.4.0
websockets==12.0
python-dateutil==2.9.0.post0
matplotlib==3.9.0
numpy==1.26.4
scipy==1.13.1
ta==0.11.0b0